Populate jwt output location for jwt filter. (#4711)
* Add output location to jwt filter config.

* Move most of model/authentication to networking/plugin/authn

* Revert change for plugin

* Remove extra call to get authn policy

* Change jwt forward to false.
