+++ pilot/pkg/networking/plugin/authn/authentication.go
@@ -15,15 +15,21 @@
 package authn
 
 import (
+	"crypto/sha1"
+	"fmt"
+
 	xdsapi "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/auth"
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
+	jwtfilter "github.com/envoyproxy/go-control-plane/envoy/config/filter/http/jwt_authn/v2alpha"
 	http_conn "github.com/envoyproxy/go-control-plane/envoy/config/filter/network/http_connection_manager/v2"
+	"github.com/gogo/protobuf/types"
 
 	authn "istio.io/api/authentication/v1alpha1"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/networking/plugin"
 	"istio.io/istio/pilot/pkg/networking/util"
+	"istio.io/istio/pkg/log"
 )
 
 const (
@@ -31,6 +37,9 @@ const (
 	// as the name defined in
 	// https://github.com/istio/proxy/blob/master/src/envoy/http/jwt_auth/http_filter_factory.cc#L50
 	jwtFilterName = "jwt-auth"
+
+	// Defautl cache duration for JWT public key. This should be moved to a global config.
+	jwtPublicKeyCacheSeconds = 60 * 5
 )
 
 // Plugin implements Istio mTLS auth
@@ -41,9 +50,114 @@ func NewPlugin() *Plugin {
 	return &Plugin{}
 }
 
+// RequireTLS returns true and pointer to mTLS params if the policy use mTLS for (peer) authentication.
+// (note that mTLS params can still be nil). Otherwise, return (false, nil).
+func RequireTLS(policy *authn.Policy) (bool, *authn.MutualTls) {
+	if policy == nil {
+		return false, nil
+	}
+	if len(policy.Peers) > 0 {
+		for _, method := range policy.Peers {
+			switch method.GetParams().(type) {
+			case *authn.PeerAuthenticationMethod_Mtls:
+				return true, method.GetMtls()
+			default:
+				continue
+			}
+		}
+	}
+	return false, nil
+}
+
+// JwksURIClusterName returns cluster name for the jwks URI. This should be used
+// to override the name for outbound cluster that are added for Jwks URI so that they
+// can be referred correctly in the JWT filter config.
+func JwksURIClusterName(hostname string, port *model.Port) string {
+	const clusterPrefix = "jwks."
+	const maxClusterNameLength = 189 - len(clusterPrefix)
+	name := hostname + "|" + port.Name
+	if len(name) > maxClusterNameLength {
+		prefix := name[:maxClusterNameLength-sha1.Size*2]
+		sum := sha1.Sum([]byte(name))
+		name = fmt.Sprintf("%s%x", prefix, sum)
+	}
+	return clusterPrefix + name
+}
+
+// CollectJwtSpecs returns a list of all JWT specs (ponters) defined the policy. This
+// provides a convenient way to iterate all Jwt specs.
+func CollectJwtSpecs(policy *authn.Policy) []*authn.Jwt {
+	ret := []*authn.Jwt{}
+	if policy == nil {
+		return ret
+	}
+	for _, method := range policy.Peers {
+		switch method.GetParams().(type) {
+		case *authn.PeerAuthenticationMethod_Jwt:
+			ret = append(ret, method.GetJwt())
+		}
+	}
+	for _, method := range policy.Origins {
+		ret = append(ret, method.Jwt)
+	}
+	return ret
+}
+
+// OutputLocationForJwtIssuer returns the header location that should be used to output payload if
+// authentication succeeds.
+func OutputLocationForJwtIssuer(issuer string) string {
+	const locationPrefix = "istio-sec-"
+	sum := sha1.Sum([]byte(issuer))
+	return locationPrefix + fmt.Sprintf("%x", sum)
+}
+
+// ConvertPolicyToJwtConfig converts policy into Jwt filter config for envoy.
+func ConvertPolicyToJwtConfig(policy *authn.Policy) *jwtfilter.JwtAuthentication {
+	policyJwts := CollectJwtSpecs(policy)
+	if len(policyJwts) == 0 {
+		return nil
+	}
+	ret := &jwtfilter.JwtAuthentication{
+		AllowMissingOrFailed: true,
+	}
+	for _, policyJwt := range policyJwts {
+		hostname, port, _, err := model.ParseJwksURI(policyJwt.JwksUri)
+		if err != nil {
+			log.Errorf("Cannot parse jwks_uri %q: %v", policyJwt.JwksUri, err)
+			continue
+		}
+
+		jwt := &jwtfilter.JwtRule{
+			Issuer:    policyJwt.Issuer,
+			Audiences: policyJwt.Audiences,
+			JwksSourceSpecifier: &jwtfilter.JwtRule_RemoteJwks{
+				RemoteJwks: &jwtfilter.RemoteJwks{
+					HttpUri: &core.HttpUri{
+						Uri: policyJwt.JwksUri,
+						HttpUpstreamType: &core.HttpUri_Cluster{
+							Cluster: JwksURIClusterName(hostname, port),
+						},
+					},
+					CacheDuration: &types.Duration{Seconds: jwtPublicKeyCacheSeconds},
+				},
+			},
+			ForwardPayloadHeader: OutputLocationForJwtIssuer(policyJwt.Issuer),
+			Forward:              false,
+		}
+		for _, location := range policyJwt.JwtHeaders {
+			jwt.FromHeaders = append(jwt.FromHeaders, &jwtfilter.JwtHeader{
+				Name: location,
+			})
+		}
+		jwt.FromParams = policyJwt.JwtParams
+		ret.Rules = append(ret.Rules, jwt)
+	}
+	return ret
+}
+
 // BuildJwtFilter returns a Jwt filter for all Jwt specs in the policy.
 func BuildJwtFilter(policy *authn.Policy) *http_conn.HttpFilter {
-	filterConfigProto := model.ConvertPolicyToJwtConfig(policy)
+	filterConfigProto := ConvertPolicyToJwtConfig(policy)
 	if filterConfigProto == nil {
 		return nil
 	}
@@ -97,7 +211,7 @@ func (*Plugin) OnOutboundCluster(env model.Environment, node model.Proxy, servic
 		return
 	}
 
-	required, _ := model.RequireTLS(model.GetConsolidateAuthenticationPolicy(mesh, config, service.Hostname, servicePort))
+	required, _ := RequireTLS(model.GetConsolidateAuthenticationPolicy(mesh, config, service.Hostname, servicePort))
 	if isDestinationExcludedForMTLS(service.Hostname, mesh.MtlsExcludedServices) || !required {
 		return
 	}
