+++ pilot/pkg/proxy/envoy/v1/authentication_test.go
@@ -62,7 +62,7 @@ func TestBuildJwksFilter(t *testing.T) {
 									"cluster": "jwks.abc.com|http",
 								},
 							},
-							"forward": true,
+							"forwardPayloadHeader": "istio-sec-da39a3ee5e6b4b0d3255bfef95601890afd80709",
 						},
 					},
 					"allowMissingOrFailed": true,
